requirements.txt
python-dotenv
requests
openai
python-docx
tavily-python